# apps/backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "VRK API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/equipment": {
            "get": {
                "description": "Returns equipment records visible inside the authenticated customer organization/session scope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "List equipment registry",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/equipment.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/equipment.EquipmentResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/equipment.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates one equipment registry record inside the authenticated customer organization and visible unit contour.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Create equipment registry record",
                "parameters": [
                    {
                        "description": "Equipment payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/equipment.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/equipment.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/equipment.EquipmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    }
                }
            }
        },
        "/equipment/{id}": {
            "get": {
                "description": "Returns one equipment record visible inside the authenticated customer organization/session scope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Get equipment registry record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/equipment.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/equipment.EquipmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates one equipment record inside the authenticated customer organization/session scope.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Update equipment registry record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Equipment patch payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/equipment.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/equipment.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/equipment.EquipmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    }
                }
            }
        },
        "/equipment/{id}/archive": {
            "post": {
                "description": "Archives one equipment record inside the authenticated customer organization/session scope.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "equipment"
                ],
                "summary": "Archive equipment registry record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Equipment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/equipment.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/equipment.EquipmentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/equipment.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "equipment.CreateRequest": {
            "type": "object",
            "properties": {
                "classification": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "documentUrl": {
                    "type": "string"
                },
                "factoryNumber": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "inventoryNumber": {
                    "type": "string"
                },
                "manufactureYear": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
        "equipment.EquipmentResponse": {
            "type": "object",
            "properties": {
                "archivedAt": {
                    "type": "string"
                },
                "classification": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "documentUrl": {
                    "type": "string"
                },
                "factoryNumber": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inventoryNumber": {
                    "type": "string"
                },
                "manufactureYear": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "measuringInstrumentCount": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unit": {
                    "$ref": "#/definitions/equipment.UnitSummary"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "equipment.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "equipment.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/equipment.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "equipment.UnitSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subdivisionId": {
                    "type": "string"
                },
                "subdivisionName": {
                    "type": "string"
                }
            }
        },
        "equipment.UpdateRequest": {
            "type": "object",
            "properties": {
                "classification": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "documentUrl": {
                    "type": "string"
                },
                "factoryNumber": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "inventoryNumber": {
                    "type": "string"
                },
                "manufactureYear": {
                    "type": "integer"
                },
                "manufacturer": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "unitId": {
                    "type": "string"
                }
            }
        },
...
    }
}